+++ app/controllers/admin_controller.rb
@@ -1,6 +1,6 @@
 class AdminController < ApplicationController
   def index 
-    @incomplete_invoice_items = InvoiceItem.incomplete_invoices
+    @incomplete_invoice_items = Invoice.incomplete_invoices
     @top_customers = Customer.top_customers
   end
 end
